Move from_config_provider to framework/
